feat(nn): promote activation calibration + salience fold into the library
The AWQ-style fold produced every published SALT number (fold alpha=0.75 on
top of the geometric ladder), but it lived in `tests/common/mod.rs` where
nothing shippable could reach it. That is why `tritium quantize` writes a
DIFFERENT configuration from the research runs: it reads a bare safetensors
file, so it cannot compute activation statistics at all. Measured cost of the
gap on SmolLM2-360M: the fold is worth 2.3x at T=2, 8.5% at T=3, 1.1% at T=4.

`fold` is model-aware, not a tensor op. It scales weight columns by the
per-channel salience and divides the PRECEDING norm by the same factor -- an
exact reparameterisation, but one that needs to know which norm feeds which
projection. So it takes an `Arch` and returns a modified one, and any command
offering calibration has to load a model rather than a file of tensors. That
is the shape of `tritium convert` (ADR 0038 WS-3); this commit is its
prerequisite.

`tests/common` now re-exports these from the library instead of keeping a
second copy. A duplicated fold is exactly the train/eval quantizer mismatch
that already forced one retraction in this repo (WS-C0).

The move is behaviour-preserving by construction: the promoted bodies diff
byte-identically against the committed originals except for two derives, and
`tape_model_conformance` -- which validates `forward` bit-exact against
`ModelRunner` -- now runs through the library copy and still passes.

`clone_arch` is deleted rather than promoted: it was a hand-written
field-for-field `Clone` and `Arch` now derives one.

Left in the test tree deliberately: `Gram`, `GramSet`, `damped_inverse`,
`calibrate_gram`. Per-group curvature allocation measured NEGATIVE -- 12.4%
lower weight SSE and 12% worse perplexity at identical bits -- so it is a
research artifact, not a shipping path.
